tpm2: move UnmarshalProtectedKeys from export_test.go to tpm2_test package
