+++ tpm2/export_test.go
@@ -20,11 +20,7 @@
 package tpm2
 
 import (
-	"errors"
-
 	"github.com/canonical/go-tpm2"
-	"golang.org/x/crypto/cryptobyte"
-	cryptobyte_asn1 "golang.org/x/crypto/cryptobyte/asn1"
 
 	"github.com/snapcore/secboot"
 )
@@ -267,27 +263,4 @@ func ValidateKeyDataFile(tpm *tpm2.TPMContext, keyFile string, authKey secboot.P
 	}
 
 	return k.Validate(tpm, authKey, session)
-}
-
-type protectedKeys struct {
-	Primary secboot.PrimaryKey
-	Unique  []byte
-}
-
-func UnmarshalProtectedKeys(data []byte) (*protectedKeys, error) {
-	s := cryptobyte.String(data)
-	if !s.ReadASN1(&s, cryptobyte_asn1.SEQUENCE) {
-		return nil, errors.New("malformed input")
-	}
-
-	pk := new(protectedKeys)
-
-	if !s.ReadASN1Bytes((*[]byte)(&pk.Primary), cryptobyte_asn1.OCTET_STRING) {
-		return nil, errors.New("malformed primary key")
-	}
-	if !s.ReadASN1Bytes(&pk.Unique, cryptobyte_asn1.OCTET_STRING) {
-		return nil, errors.New("malformed unique key")
-	}
-
-	return pk, nil
 }
